requirements.txt
certifi==2020.12.5
chardet==4.0.0
idna==2.10
pyTelegramBotAPI==3.7.9
python-decouple==3.4
requests==2.25.1
urllib3==1.26.4
